requirements.txt
appdirs==1.4.4
autopep8==1.5.7
black==21.6b0
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
flake8==3.9.2
idna==3.2
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.0
pandas==1.3.0
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
requests==2.26.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webdriver-manager==3.4.2
